20220807-1043 - Updated packages of the system - Created the homepage with planned functionality of the system: -- Animal List -- Guess the animal -- Have animal guessed -- About page
