+++ .packages
@@ -3,7 +3,7 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2022-08-07 11:31:58.158341.
+# Generated by pub on 2022-08-07 13:18:08.851420.
 async:file:///C:/creations/third-party-apps/flutter_windows_3.0.5-stable/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
 boolean_selector:file:///C:/creations/third-party-apps/flutter_windows_3.0.5-stable/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
 characters:file:///C:/creations/third-party-apps/flutter_windows_3.0.5-stable/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
