+++ android/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Aug 07 12:36:35 CDT 2022
+gradle.version=7.4
