+++ android/app/src/main/kotlin/com/mythcreaturescreations/twenty_questions_na_mammals/MainActivity.kt
@@ -1,6 +1,1 @@
 package com.mythcreaturescreations.twenty_questions_na_mammals
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
