+++ android/local.properties
@@ -1,2 +1,5 @@
-sdk.dir=C:\\Users\\mythc\\AppData\\Local\\Android\\sdk
-flutter.sdk=C:\\creations\\third-party-apps\\flutter_windows_3.0.5-stable\\flutter
+sdk.dir=C\:\\Users\\mythc\\AppData\\Local\\Android\\sdk
+flutter.sdk=C\:\\creations\\third-party-apps\\flutter_windows_3.0.5-stable\\flutter
+flutter.buildMode=debug
+flutter.versionName=1.0.0
+flutter.versionCode=1
