+++ local.properties
@@ -1,0 +1,8 @@
+## This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+#
+# Location of the SDK. This is only used by Gradle.
+# For customization when using a Version Control System, please read the
+# header note.
+#Sun Aug 07 12:41:32 CDT 2022
+sdk.dir=C\:\\Users\\mythc\\AppData\\Local\\Android\\Sdk
